fix: Исправлены замечания по комментариям
